fix: Correct telemetry.IncrCounter type mismatch in bank keeper

+++ x/bank/keeper/send.go
@@ -198,7 +198,7 @@ func (k BaseSendKeeper) InputOutputCoins(ctx context.Context, input types.Input,
 		// such as delegated fee messages.
 		accExists := k.ak.HasAccount(ctx, updatedAddressBz)
 		if !accExists {
-			defer telemetry.IncrCounter(1, "new", "account") //nolint:staticcheck // TODO: switch to OpenTelemetry
+			defer telemetry.IncrCounter(float32(1), "new", "account") //nolint:staticcheck // TODO: switch to OpenTelemetry
 			k.ak.SetAccount(ctx, k.ak.NewAccountWithAddress(ctx, updatedAddressBz))
 		}
 	}
@@ -261,7 +261,7 @@ func (k BaseSendKeeper) ensureAccountCreated(ctx context.Context, toAddr sdk.Acc
 	// such as delegated fee messages.
 	accExists := k.ak.HasAccount(ctx, toAddr)
 	if !accExists {
-		defer telemetry.IncrCounter(1, "new", "account") //nolint:staticcheck // TODO: switch to OpenTelemetry
+		defer telemetry.IncrCounter(float32(1), "new", "account") //nolint:staticcheck // TODO: switch to OpenTelemetry
 		k.ak.SetAccount(ctx, k.ak.NewAccountWithAddress(ctx, toAddr))
 	}
 }
